backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
Flask-SQLAlchemy>=3.0.3
pymysql==1.1.0
cryptography==43.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==20.1.0
Werkzeug==2.3.0
authlib==1.3.0
PyJWT==2.8.0
requests==2.32.3
python-jose[cryptography]
python-jose[cryptography]==3.3.0
python-dateutil


# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
coverage==7.2.7